code_base/v3_code_base/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-assistant-v3"
version = "0.1.0"
description = "Modular multimodal alignment and instruction tuning stack"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "transformers",
    "datasets",
    "tqdm",
    "soundfile",
    "requests",
    "pyyaml",
]

[tool.setuptools.packages.find]
where = ["src"]
